+++ CHANGELOG.md
@@ -8,12 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `RawOsError` type alias [#739]
+- `SysRng` behind new feature `sys_rng` [#751]
 
 ### Changed
 - Use Edition 2024 and MSRV 1.85 [#749]
 
 [#739]: https://github.com/rust-random/getrandom/pull/739
 [#749]: https://github.com/rust-random/getrandom/pull/749
+[#751]: https://github.com/rust-random/getrandom/pull/751
 
 ## [0.3.4] - 2025-10-14